// src/projects/model/epic.ts
export class Epic {
  Title?: string;
  Status?: string;
  Assignees?: string;
  Labels?: string[];
  Type?: string;
  Sprint?: string;
  SprintPoints?: number;
  Release?: string;
  Repository?: string;
  Period?: string;
  Partner?: string;
  EpicPoints?: number;
  EpicPointsDone?: number;
  EpicPointsRemaining?: number;
}